+++ xarray/datatree_/docs/source/whats-new.rst
@@ -39,6 +39,8 @@ Deprecations
 Bug fixes
 ~~~~~~~~~
 
+- Fix unittests on i386. (:pull:`249`)
+  By `Antonio Valentino <https://github.com/avalentino>`_.
 - Ensure nodepath class is compatible with python 3.12 (:pull:`260`)
   By `Max Grover <https://github.com/mgrover1>`_.